clk_pll_16fft.c: 6: yes: MISRA.ETYPE.COMP.ASSIGN.2012
Rule 10.6: The value of a composite expression shall not be assigned to an
object with wider essential type - 19 violations fixed (IDs: 1045, 1046,
1047, 1048, 1049, 1050, 1051, 1052, 1054, 1055, 1056, 1057, 1058, 1059,
1060, 1061, 1062, 3642, 3643). Removed unnecessary casts from readl()
return values and composite expressions to prevent assignment of narrower
essential type to wider essential type variables.

+++ plat/ti/k3low/common/drivers/pm/clock/clk_pll_16fft.c
@@ -310,8 +310,8 @@ static bool clk_pll_16fft_check_lock(const struct clk_data_pll_16fft *pll)
 {
 	uint32_t stat;
 
-	stat = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_STAT(pll->idx));
-	return ((uint32_t)(stat & PLL_16FFT_STAT_LOCK)) != 0U;
+	stat = readl(pll->base + (uint32_t) PLL_16FFT_STAT(pll->idx));
+	return (stat & PLL_16FFT_STAT_LOCK) != 0U;
 }
 
 static int32_t clk_pll_16fft_enable_pll(const struct clk_data_pll_16fft *pll)
@@ -319,9 +319,9 @@ static int32_t clk_pll_16fft_enable_pll(const struct clk_data_pll_16fft *pll)
 	uint32_t ctrl;
 	int32_t err = SUCCESS;
 
-	ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
+	ctrl = readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
 
-	if (((uint32_t)(ctrl & PLL_16FFT_CTRL_PLL_EN)) == 0U) {
+	if ((ctrl & PLL_16FFT_CTRL_PLL_EN) == 0U) {
 		ctrl |= PLL_16FFT_CTRL_PLL_EN;
 		ti_clk_writel(ctrl, pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
 		osal_delay(1UL); /* Wait 1us */
@@ -335,9 +335,9 @@ static int32_t clk_pll_16fft_disable_pll(const struct clk_data_pll_16fft *pll)
 	uint32_t ctrl;
 	int32_t err = SUCCESS;
 
-	ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
+	ctrl = readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
 
-	if (((uint32_t)(ctrl & PLL_16FFT_CTRL_PLL_EN)) != 0U) {
+	if ((ctrl & PLL_16FFT_CTRL_PLL_EN) != 0U) {
 		ctrl &= (uint32_t)~PLL_16FFT_CTRL_PLL_EN;
 		ti_clk_writel(ctrl, pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
 		osal_delay(1UL); /* Wait 1us */
@@ -490,8 +490,8 @@ static bool clk_pll_16fft_is_bypass(const struct clk_data_pll_16fft *pll)
 	uint32_t ctrl;
 
 	/* IDLE Bypass */
-	ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
-	ret = ((uint32_t)(ctrl & PLL_16FFT_CTRL_BYPASS_EN)) != 0U;
+	ctrl = readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
+	ret = (ctrl & PLL_16FFT_CTRL_BYPASS_EN) != 0U;
 
 	return ret;
 }
@@ -509,7 +509,7 @@ static int32_t clk_pll_16fft_bypass(struct clk *clock_ptr, bool bypass)
 	pll = container_of(data_pll, const struct clk_data_pll_16fft,
 			   data_pll);
 
-	ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
+	ctrl = readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
 	if (bypass) {
 		/* Enable bypass */
 		ctrl |= PLL_16FFT_CTRL_BYPASS_EN;
@@ -539,9 +539,9 @@ static uint32_t clk_pll_16fft_get_freq_internal(struct clk *clock_ptr,
 	data_pll = container_of(clock_data->data, const struct clk_data_pll,
 				data);
 	pll = container_of(data_pll, const struct clk_data_pll_16fft, data_pll);
-	freq_ctrl0 = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL0(pll->idx));
-	freq_ctrl1 = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL1(pll->idx));
-	div_ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
+	freq_ctrl0 = readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL0(pll->idx));
+	freq_ctrl1 = readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL1(pll->idx));
+	div_ctrl = readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
 
 	pllm = (uint32_t)(freq_ctrl0 & PLL_16FFT_FREQ_CTRL0_FB_DIV_INT_MSK);
 	pllm >>= PLL_16FFT_FREQ_CTRL0_FB_DIV_INT_SHIFT;
@@ -624,13 +624,13 @@ static bool clk_pll_16fft_program_freq(struct clk *pll_clk,
 	/* Disable the PLL */
 	(void)clk_pll_16fft_disable_pll(pll);
 
-	cfg = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_CFG(pll->idx));
-	pll_type = (uint32_t)((cfg & PLL_16FFT_CFG_PLL_TYPE_MASK) >> PLL_16FFT_CFG_PLL_TYPE_SHIFT);
+	cfg = readl(pll->base + (uint32_t) PLL_16FFT_CFG(pll->idx));
+	pll_type = (cfg & PLL_16FFT_CFG_PLL_TYPE_MASK) >> PLL_16FFT_CFG_PLL_TYPE_SHIFT;
 
 	/* Program the new rate */
-	freq_ctrl0 = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL0(pll->idx));
-	freq_ctrl1 = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL1(pll->idx));
-	div_ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
+	freq_ctrl0 = readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL0(pll->idx));
+	freq_ctrl1 = readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL1(pll->idx));
+	div_ctrl = readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
 
 	freq_ctrl0 &= (uint32_t)~PLL_16FFT_FREQ_CTRL0_FB_DIV_INT_MSK;
 	freq_ctrl0 |= (uint32_t)(pllm << PLL_16FFT_FREQ_CTRL0_FB_DIV_INT_SHIFT);
@@ -642,7 +642,7 @@ static bool clk_pll_16fft_program_freq(struct clk *pll_clk,
 	div_ctrl |= (uint32_t)(plld << PLL_16FFT_DIV_CTRL_REF_DIV_SHIFT);
 
 	/* Make sure we have fractional support if required */
-	ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
+	ctrl = readl(pll->base + (uint32_t) PLL_16FFT_CTRL(pll->idx));
 
 	/* Don't use internal bypass,it is not glitch free. Always prefer glitchless bypass */
 	ctrl &= (uint32_t)~PLL_16FFT_CTRL_INTL_BYP_EN;
@@ -764,9 +764,9 @@ static uint32_t clk_pll_16fft_internal_set_freq(struct clk *pll_clk,
 
 	was_bypass = clk_pll_16fft_is_bypass(pll);
 
-	freq_ctrl0 = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL0(pll->idx));
-	freq_ctrl1 = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL1(pll->idx));
-	div_ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
+	freq_ctrl0 = readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL0(pll->idx));
+	freq_ctrl1 = readl(pll->base + (uint32_t) PLL_16FFT_FREQ_CTRL1(pll->idx));
+	div_ctrl = readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
 
 	/* Check current values */
 	prev_pllm = (uint32_t)(freq_ctrl0 & PLL_16FFT_FREQ_CTRL0_FB_DIV_INT_MSK);
@@ -867,7 +867,7 @@ static uint32_t clk_pll_16fft_internal_set_freq_from_pll_table(struct clk *pll_c
 		 *
 		 * Note: We break up the calculation in order to avoid a div64.
 		 */
-		clkod_plld = soc_pll_table[data_pll->pll_entries[i]].plld *
+		clkod_plld = (uint32_t)soc_pll_table[data_pll->pll_entries[i]].plld *
 			     soc_pll_table[data_pll->pll_entries[i]].clkod;
 
 			actual64 = ((uint64_t) input / clkod_plld) *
@@ -1403,7 +1403,7 @@ static bool clk_pll_16fft_postdiv_set_div(struct clk *clock_ptr, uint32_t d)
 		pll = container_of(data_pll, const struct clk_data_pll_16fft,
 				   data_pll);
 
-		div_ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
+		div_ctrl = readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
 		div_ctrl &= (uint32_t)~PLL_16FFT_DIV_CTRL_POST_DIV2_MASK;
 		div_ctrl &= (uint32_t)~PLL_16FFT_DIV_CTRL_POST_DIV1_MASK;
 
@@ -1444,7 +1444,7 @@ static uint32_t clk_pll_16fft_postdiv_get_div(struct clk *clock_ptr)
 		pll = container_of(data_pll,
 				   const struct clk_data_pll_16fft, data_pll);
 
-		div_ctrl = (uint32_t)readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
+		div_ctrl = readl(pll->base + (uint32_t) PLL_16FFT_DIV_CTRL(pll->idx));
 		post_div1 = (uint32_t)(div_ctrl & PLL_16FFT_DIV_CTRL_POST_DIV1_MASK);
 		post_div1 >>= PLL_16FFT_DIV_CTRL_POST_DIV1_SHIFT;
 		post_div2 = div_ctrl & PLL_16FFT_DIV_CTRL_POST_DIV2_MASK;
